fix AttrVar
修复文本转long long的bug

+++ Dice/DiceAttrVar.cpp
@@ -149,7 +149,7 @@ long long AttrVar::to_ll()const {
 		break;
 	case AttrType::Text:
 		try {
-			return stoi(text);
+			return stoll(text);
 		}
 		catch (...) {
 			return 0;
